requirements.txt
Flask==2.3.3
requests==2.31.0
pyjwt==2.8.0
cryptography==41.0.4
protobuf==4.24.3
pycryptodome==3.18.0
psutil==5.9.5
urllib3==1.26.16
gunicorn==21.2.0